Move schemas around to avoid circular references
