+++ docs/documentation/type._editables.*.code_block_options.language.type.yml
@@ -1,0 +1,6 @@
+gid: type._editables.*.code_block_options.language.type
+url: /configuration-file/types/_editables/*/code_block_options/language/type/
+title: ''
+description: Sets an input type, which controls how this input appears and behaves.
+examples: []
+show_in_navigation: false
