+++ docs/documentation/type._editables.*.image_options.alt.options.icon.yml
@@ -1,0 +1,6 @@
+gid: type._editables.*.image_options.alt.options.icon
+url: /configuration-file/types/_editables/*/image_options/alt/options/icon/
+title: ''
+description: Icon shown beside the input.
+examples: []
+show_in_navigation: false
