+++ docs/documentation/type._editables.*.image_options.alt.options.icon_color.yml
@@ -1,0 +1,6 @@
+gid: type._editables.*.image_options.alt.options.icon_color
+url: /configuration-file/types/_editables/*/image_options/alt/options/icon_color/
+title: ''
+description: ''
+examples: []
+show_in_navigation: false
