+++ docs/documentation/type._editables.*.image_options.alt.options.placeholder.yml
@@ -1,0 +1,6 @@
+gid: type._editables.*.image_options.alt.options.placeholder
+url: /configuration-file/types/_editables/*/image_options/alt/options/placeholder/
+title: ''
+description: Text shown when this input has no value.
+examples: []
+show_in_navigation: false
