+++ docs/documentation/type._editables.*.image_options.alt.options.yml
@@ -1,0 +1,6 @@
+gid: type._editables.*.image_options.alt.options
+url: /configuration-file/types/_editables/*/image_options/alt/options/
+title: ''
+description: Options that are specific to Text Inputs.
+examples: []
+show_in_navigation: false
