+++ docs/documentation/type._editables.*.image_options.src.options.hide_link_to_page.yml
@@ -1,0 +1,9 @@
+gid: type._editables.*.image_options.src.options.hide_link_to_page
+url: >-
+  /configuration-file/types/_editables/*/image_options/src/options/hide_link_to_page/
+title: ''
+description: >-
+  Hides the option to link to a page. This does not prevent typing a file's
+  output URL in the input.
+examples: []
+show_in_navigation: false
