+++ docs/documentation/type._editables.*.image_options.src.options.yml
@@ -1,0 +1,6 @@
+gid: type._editables.*.image_options.src.options
+url: /configuration-file/types/_editables/*/image_options/src/options/
+title: ''
+description: Options that are specific to URL Inputs.
+examples: []
+show_in_navigation: false
