+++ docs/documentation/type._editables.*.image_options.src.yml
@@ -1,0 +1,8 @@
+gid: type._editables.*.image_options.src
+url: /configuration-file/types/_editables/*/image_options/src/
+title: ''
+description: >-
+  Provides an editing interface for relative, absolute, and fully qualified
+  URLs.
+examples: []
+show_in_navigation: false
