+++ docs/documentation/type._editables.*.image_options.title.options.icon.yml
@@ -1,0 +1,6 @@
+gid: type._editables.*.image_options.title.options.icon
+url: /configuration-file/types/_editables/*/image_options/title/options/icon/
+title: ''
+description: Icon shown beside the input.
+examples: []
+show_in_navigation: false
