+++ docs/documentation/type._editables.*.image_options.title.options.placeholder.yml
@@ -1,0 +1,7 @@
+gid: type._editables.*.image_options.title.options.placeholder
+url: >-
+  /configuration-file/types/_editables/*/image_options/title/options/placeholder/
+title: ''
+description: Text shown when this input has no value.
+examples: []
+show_in_navigation: false
