+++ docs/documentation/type._editables.*.link_options.href.options.hide_link_to_email_address.yml
@@ -1,0 +1,9 @@
+gid: type._editables.*.link_options.href.options.hide_link_to_email_address
+url: >-
+  /configuration-file/types/_editables/*/link_options/href/options/hide_link_to_email_address/
+title: ''
+description: >-
+  Hides the option to link to an email address. This does not prevent typing a
+  `mailto:` link in the input.
+examples: []
+show_in_navigation: false
