+++ docs/documentation/type._editables.*.link_options.href.options.hide_link_to_file.yml
@@ -1,0 +1,9 @@
+gid: type._editables.*.link_options.href.options.hide_link_to_file
+url: >-
+  /configuration-file/types/_editables/*/link_options/href/options/hide_link_to_file/
+title: ''
+description: >-
+  Hides the options to link to an existing file, and upload a new file. This
+  does not prevent typing a file path in the input.
+examples: []
+show_in_navigation: false
