+++ docs/documentation/type._editables.*.link_options.href.options.hide_link_to_telephone.yml
@@ -1,0 +1,9 @@
+gid: type._editables.*.link_options.href.options.hide_link_to_telephone
+url: >-
+  /configuration-file/types/_editables/*/link_options/href/options/hide_link_to_telephone/
+title: ''
+description: >-
+  Hides the option to link to a telephone number. This does not prevent typing a
+  `tel:` link in the input.
+examples: []
+show_in_navigation: false
