+++ docs/documentation/type._editables.*.link_options.href.options.yml
@@ -1,0 +1,6 @@
+gid: type._editables.*.link_options.href.options
+url: /configuration-file/types/_editables/*/link_options/href/options/
+title: ''
+description: Options that are specific to URL Inputs.
+examples: []
+show_in_navigation: false
