+++ docs/documentation/type._editables.*.link_options.rel_nofollow.yml
@@ -1,0 +1,6 @@
+gid: type._editables.*.link_options.rel_nofollow
+url: /configuration-file/types/_editables/*/link_options/rel_nofollow/
+title: ''
+description: Provides an editing interface for true or false values.
+examples: []
+show_in_navigation: false
