+++ docs/documentation/type._editables.*.link_options.title.options.yml
@@ -1,0 +1,6 @@
+gid: type._editables.*.link_options.title.options
+url: /configuration-file/types/_editables/*/link_options/title/options/
+title: ''
+description: Options that are specific to Text Inputs.
+examples: []
+show_in_navigation: false
