+++ docs/documentation/type._editables.*.link_options.title.type.yml
@@ -1,0 +1,6 @@
+gid: type._editables.*.link_options.title.type
+url: /configuration-file/types/_editables/*/link_options/title/type/
+title: ''
+description: Sets an input type, which controls how this input appears and behaves.
+examples: []
+show_in_navigation: false
