+++ src/editables.ts
@@ -1,7 +1,7 @@
 import * as z from 'zod';
 import { ImageOptionsSchema } from './image-options.ts';
 import { PathsSchema } from './paths.ts';
-import { BooleanInputSchema, SelectInputSchema, TextInputSchema, UrlInputSchema } from './inputs.ts';
+import { BooleanInputSchema, SelectInputSchema, TextInputSchema, UrlInputSchema } from './input-base.ts';
 
 export const TextEditableSchema = z.object({
 	paths: PathsSchema.optional(),
@@ -27,16 +27,16 @@ export const TextEditableSchema = z.object({
 		description: 'Enables a control to create hyperlinks around selected text.',
 	}),
 
-	// link_options: z.object({
-	// 	get href() { return UrlInputSchema.optional() },
-	// 	get title() { return TextInputSchema.optional() },
-	// 	get is_target_blank() { return BooleanInputSchema.optional() },
-	// 	get rel_nofollow() { return BooleanInputSchema.optional() },
-	// 	get rel_noreferrer() { return BooleanInputSchema.optional() }
-	// }).optional().meta({
-	// 	id: 'type._editables.*.link_options',
-	// 	description: 'Configure the controls to edit hyperlinks around selected text.',
-	// }),
+	link_options: z.object({
+		get href() { return UrlInputSchema.optional() },
+		get title() { return TextInputSchema.optional() },
+		get is_target_blank() { return BooleanInputSchema.optional() },
+		get rel_nofollow() { return BooleanInputSchema.optional() },
+		get rel_noreferrer() { return BooleanInputSchema.optional() }
+	}).optional().meta({
+		id: 'type._editables.*.link_options',
+		description: 'Configure the controls to edit hyperlinks around selected text.',
+	}),
 
 	redo: z.boolean().default(false).optional().meta({
 		id: 'type._editables.*.redo',
@@ -120,12 +120,12 @@ export const ToolbarOptionsSchema = z.object({
 		description: 'Enables a control to insert a code block.',
 	}),
 
-	// code_block_options: z.object({
-	// 	get language() { return SelectInputSchema.optional() },
-	// }).optional().meta({
-	// 	id: 'type._editables.*.code_block_options',
-	// 	description: 'Configure the controls to edit a code block.',
-	// }),
+	code_block_options: z.object({
+		get language() { return SelectInputSchema.optional() },
+	}).optional().meta({
+		id: 'type._editables.*.code_block_options',
+		description: 'Configure the controls to edit a code block.',
+	}),
 
 	code: z.boolean().default(false).optional().meta({
 		id: 'type._editables.*.code',
@@ -158,11 +158,10 @@ export const ToolbarOptionsSchema = z.object({
 	get image_options() {
 		return (
 			z.object({
-				// src: UrlInputSchema.optional(),
-				// alt: TextInputSchema.optional(),
-				// title: TextInputSchema.optional(),
-				loading: BooleanInputSchema.optional(),
-				test: z.string(),
+				src: UrlInputSchema.optional(),
+				alt: TextInputSchema.optional(),
+				title: TextInputSchema.optional(),
+				loading: BooleanInputSchema.optional()
 			}).optional().meta({
 				id: 'type._editables.*.image_options',
 				description: 'Configure the controls to edit an image.',
